Add individual topping add functions and modify
The add functions from cheese, protein, and veggie takes a list of items
and checks to see if it is already included in the list of toppings. If
it is a new topping, it will add the topping.

The modify function takesin the property to be modified, and the new
value for the property.