Fix undefined error and change Pizza constructor
Was getting an undefined error in the forEach callback of the add
toppings methods. Defining this.toppings as a variable before the
forEach callback solved the problem.

Change the Pizza constructor to make it easier to define a pizza and
it's toppings when you create the instance of the new Pizza by adding
properties to the constructor.